wrap wasmi_core::FuncType in wasmi
